refactor: rename package handler
