+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handler/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe.handlermapping;
+package nextstep.mvc.tobe.handler;
 
 import nextstep.web.annotation.RequestMethod;
