+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handler/HandlerMapping.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe.handlermapping;
+package nextstep.mvc.tobe.handler;
 
 import javax.servlet.http.HttpServletRequest;
