+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -1,8 +1,8 @@
 package nextstep.mvc.tobe;
 
 import nextstep.db.DataBase;
-import nextstep.mvc.tobe.handlermapping.AnnotationHandlerMapping;
-import nextstep.mvc.tobe.handlermapping.HandlerExecution;
+import nextstep.mvc.tobe.handler.AnnotationHandlerMapping;
+import nextstep.mvc.tobe.handler.HandlerExecution;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
